# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "description": "APIs for Scoial Networking Platform",
    "title": "Scoial Networking Platform APIs (Postaway - II)"
  },
  "servers": [
    {
      "url": "http://localhost:2500"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "paths": {
...
    "/api/posts": {
      "get": {
        "tags": ["Post"],
        "summary": "Get all posts of user",
        "description": "Retrieve all posts for a specific user to display on their profile page.",
        "security": [
          {
            "JWT": {}
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Post not found."
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },

      "post": {
        "tags": ["Post"],
        "summary": " Create a new post",
        "description": "Create a new post with an image upload functionality.",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "caption": {
                    "type": "string"
                  },
                  "imageUrl": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": {}
          }
        ],
        "responses": {
          "201": {
            "description": "Post Created."
          },
          "400": {
            "description": "Something went wrong."
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}